+++ app/src/main/java/com/paw/key/domain/usecase/GetDbtiQuestionsUseCase.kt
@@ -1,7 +1,7 @@
 package com.paw.key.domain.usecase
 
 import com.paw.key.domain.entity.dbti.DbtiQuestionEntity
-import com.paw.key.domain.repository.DbtiRepository
+import com.paw.key.domain.repository.DBTI.DbtiRepository
 import javax.inject.Inject
 
 class GetDbtiQuestionsUseCase @Inject constructor(
