+++ app/src/main/java/com/paw/key/domain/usecase/SubmitDbtiResultUseCase.kt
@@ -1,7 +1,7 @@
 package com.paw.key.domain.usecase
 
 import com.paw.key.domain.entity.DBTI.DbtiResultEntity
-import com.paw.key.domain.repository.DbtiRepository
+import com.paw.key.domain.repository.DBTI.DbtiRepository
 import javax.inject.Inject
 
 class SubmitDbtiResultUseCase @Inject constructor(
